Move Responses AdditionalProperties test into responses_test.go
Address review feedback: consolidate into the package's canonical test file.
